+++ noodles-fasta/src/sequence.rs
@@ -1,0 +1,5 @@
+//! Sequence format.
+
+pub mod record;
+
+pub use self::record::Record;